Create utils folder, move shared attributes
