+++ assets/js/utils/shared-attributes.js
